RSM reduce diffs
