+++ firmware/quadruna/RSM/src/io/io_fans.h
@@ -1,9 +1,13 @@
 #pragma once
 
+#include <stdbool.h>
+#include "app_utils.h"
+
 #ifdef TARGET_EMBEDDED
 #include "hw_gpio.h"
-void io_fans_init(const Gpio *acc_fan_in, const Gpio *rad_fan_in);
+void io_fan_init(const Gpio *acc_fan_in, const Gpio *rad_fan_in);
 #endif
 
-void io_fans_accfan_set(bool val);
-void io_fans_radfan_set(bool val);
+void io_acc_fan_set(bool on);
+
+void io_rad_fan_set(bool on);
